# client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nodriver scraping service",
    "description": "Command-driven browser automation over nodriver (undetected Chrome). Run typed scenarios against bot-protected sites such as Ozon.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/run": {
      "post": {
        "tags": [
          "scrape"
        ],
        "summary": "Run",
        "description": "Execute a scenario (list of steps) in a real Chrome and return the\nextracted data. Use `start_url` for the initial navigation.",
        "operationId": "run_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ozon/price": {
      "post": {
        "tags": [
          "ozon"
        ],
        "summary": "Ozon Price",
        "description": "Convenience endpoint: open an Ozon product page and return its price.",
        "operationId": "ozon_price_ozon_price_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OzonPriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OzonPriceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClickStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "click",
            "title": "Action",
            "default": "click"
          },
          "selector": {
            "type": "string",
            "title": "Selector",
            "description": "CSS selector of element to click."
          },
          "timeout_s": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Timeout S",
            "default": 15.0
          }
        },
        "type": "object",
        "required": [
          "selector"
        ],
        "title": "ClickStep"
      },
      "Cookie": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Cookie name."
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "Cookie value."
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain",
            "description": "Cookie domain, e.g. '.ozon.ru'. Defaults to the registrable domain of the first navigated URL.",
            "examples": [
              ".ozon.ru"
            ]
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Cookie path.",
            "default": "/"
          }
        },
        "type": "object",
        "required": [
          "name",
          "value"
        ],
        "title": "Cookie",
        "description": "A cookie set before the first navigation. Useful for pinning a\nmarketplace storefront to a region (e.g. Ozon/WB delivery region) so the\nprices and availability match what a buyer in that region sees."
      },
      "EvalStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "eval",
            "title": "Action",
            "default": "eval"
          },
          "expression": {
            "type": "string",
            "title": "Expression",
            "description": "JS expression; its value is captured."
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Key to store the result under (omit to discard)."
          }
        },
        "type": "object",
        "required": [
          "expression"
        ],
        "title": "EvalStep",
        "description": "Run arbitrary JavaScript in the page and capture the returned value."
      },
      "ExtractStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "extract",
            "title": "Action",
            "default": "extract"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Key under which the result is returned."
          },
          "selector": {
            "type": "string",
            "title": "Selector",
            "description": "CSS selector."
          },
          "kind": {
            "type": "string",
            "enum": [
              "text",
              "html",
              "attr"
            ],
            "title": "Kind",
            "description": "What to read: visible text, innerHTML, or an attribute.",
            "default": "text"
          },
          "attr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attr",
            "description": "Attribute name (required when kind='attr')."
          },
          "many": {
            "type": "boolean",
            "title": "Many",
            "description": "If true, returns a list for all matches instead of the first.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "selector"
        ],
        "title": "ExtractStep",
        "description": "Extract values from the DOM via a CSS selector. Result is stored in\nthe response `data` under `name`."
      },
      "FindTextStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "find_text",
            "title": "Action",
            "default": "find_text"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Key under which the result is returned."
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Text to search for."
          },
          "timeout_s": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Timeout S",
            "default": 10.0
          }
        },
        "type": "object",
        "required": [
          "name",
          "text"
        ],
        "title": "FindTextStep",
        "description": "Find an element by its text (nodriver's smart text search) and store\nthe enclosing element's text."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status",
            "default": "ok"
          },
          "browser_ready": {
            "type": "boolean",
            "title": "Browser Ready"
          },
          "chrome_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chrome Path"
          }
        },
        "type": "object",
        "required": [
          "browser_ready"
        ],
        "title": "HealthResponse"
      },
      "NavigateStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "navigate",
            "title": "Action",
            "default": "navigate"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "URL to open."
          },
          "new_tab": {
            "type": "boolean",
            "title": "New Tab",
            "description": "Open in a new tab instead of reusing.",
            "default": false
          },
          "settle_seconds": {
            "type": "number",
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Settle Seconds",
            "description": "Seconds to wait after load for JS to settle.",
            "default": 3.0
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "NavigateStep"
      },
      "OzonPriceRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Ozon product page URL."
          },
          "proxy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Proxy"
              },
              {
                "type": "null"
              }
            ]
          },
          "cookies": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Cookie"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cookies",
            "description": "Cookies set before navigation (e.g. region pinning)."
          },
          "headless": {
            "type": "boolean",
            "title": "Headless",
            "description": "Keep false; Ozon blocks headless.",
            "default": false
          },
          "use_cache": {
            "type": "boolean",
            "title": "Use Cache",
            "description": "Serve a recent cached value if present.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "OzonPriceRequest"
      },
      "OzonPriceResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "price_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Text",
            "description": "Raw price block text as shown on the page."
          },
          "price_value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Value",
            "description": "Best-effort main price parsed to an integer (rubles)."
          },
          "card_price_value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Price Value",
            "description": "Best-effort 'Ozon Card' price parsed to an integer."
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At",
            "description": "ISO timestamp of fetch."
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "url"
        ],
        "title": "OzonPriceResponse"
      },
      "Proxy": {
        "properties": {
          "server": {
            "type": "string",
            "title": "Server",
            "description": "Proxy address, e.g. 'http://1.2.3.4:8080' or '1.2.3.4:8080'.",
            "examples": [
              "http://1.2.3.4:8080"
            ]
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "Proxy username (optional)."
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "Proxy password (optional)."
          }
        },
        "type": "object",
        "required": [
          "server"
        ],
        "title": "Proxy",
        "description": "A single proxy. Authenticated proxies are handled via CDP Fetch,\nso user/pass do NOT need to be embedded in the server URL."
      },
      "RunRequest": {
        "properties": {
          "steps": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/NavigateStep"
                },
                {
                  "$ref": "#/components/schemas/WaitForStep"
                },
                {
                  "$ref": "#/components/schemas/WaitForAnyStep"
                },
                {
                  "$ref": "#/components/schemas/WaitForTextStep"
                },
                {
                  "$ref": "#/components/schemas/SleepStep"
                },
                {
                  "$ref": "#/components/schemas/ClickStep"
                },
                {
                  "$ref": "#/components/schemas/ScrollStep"
                },
                {
                  "$ref": "#/components/schemas/ExtractStep"
                },
                {
                  "$ref": "#/components/schemas/FindTextStep"
                },
                {
                  "$ref": "#/components/schemas/EvalStep"
                },
                {
                  "$ref": "#/components/schemas/ScreenshotStep"
                }
              ],
              "discriminator": {
                "propertyName": "action",
                "mapping": {
                  "click": "#/components/schemas/ClickStep",
                  "eval": "#/components/schemas/EvalStep",
                  "extract": "#/components/schemas/ExtractStep",
                  "find_text": "#/components/schemas/FindTextStep",
                  "navigate": "#/components/schemas/NavigateStep",
                  "screenshot": "#/components/schemas/ScreenshotStep",
                  "scroll": "#/components/schemas/ScrollStep",
                  "sleep": "#/components/schemas/SleepStep",
                  "wait_for": "#/components/schemas/WaitForStep",
                  "wait_for_any": "#/components/schemas/WaitForAnyStep",
                  "wait_for_text": "#/components/schemas/WaitForTextStep"
                }
              }
            },
            "type": "array",
            "minItems": 1,
            "title": "Steps",
            "description": "Ordered list of steps."
          },
          "start_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Url",
            "description": "Convenience: navigate here before running steps (equivalent to a leading navigate step)."
          },
          "proxy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Proxy"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional proxy for this run."
          },
          "cookies": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Cookie"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cookies",
            "description": "Cookies set before the first navigation (e.g. region pinning). Domain defaults to the start_url's registrable domain."
          },
          "headless": {
            "type": "boolean",
            "title": "Headless",
            "description": "Headless Chrome. NOTE: Ozon blocks headless — keep this false.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "RunRequest",
        "examples": [
          {
            "start_url": "https://www.ozon.ru/product/example-123456/",
            "steps": [
              {
                "action": "wait_for",
                "selector": "[data-widget=webPrice]"
              },
              {
                "action": "extract",
                "kind": "text",
                "name": "price",
                "selector": "[data-widget=webPrice]"
              },
              {
                "action": "extract",
                "name": "title",
                "selector": "h1"
              }
            ]
          }
        ]
      },
      "RunResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "final_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Url"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "Extracted values keyed by step `name`."
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepResult"
            },
            "type": "array",
            "title": "Steps"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "elapsed_ms"
        ],
        "title": "RunResponse"
      },
      "ScreenshotStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "screenshot",
            "title": "Action",
            "default": "screenshot"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Key under which the base64 image is returned."
          },
          "full_page": {
            "type": "boolean",
            "title": "Full Page",
            "description": "Capture the full page vs viewport.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ScreenshotStep"
      },
      "ScrollStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "scroll",
            "title": "Action",
            "default": "scroll"
          },
          "direction": {
            "type": "string",
            "enum": [
              "down",
              "up"
            ],
            "title": "Direction",
            "default": "down"
          },
          "amount": {
            "type": "integer",
            "title": "Amount",
            "description": "Scroll amount (percent of viewport) per step.",
            "default": 50
          },
          "times": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Times",
            "description": "How many times to scroll.",
            "default": 1
          }
        },
        "type": "object",
        "title": "ScrollStep"
      },
      "SleepStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "sleep",
            "title": "Action",
            "default": "sleep"
          },
          "seconds": {
            "type": "number",
            "maximum": 60.0,
            "exclusiveMinimum": 0.0,
            "title": "Seconds"
          }
        },
        "type": "object",
        "required": [
          "seconds"
        ],
        "title": "SleepStep"
      },
      "StepResult": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "index",
          "action",
          "ok"
        ],
        "title": "StepResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WaitForAnyStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "wait_for_any",
            "title": "Action",
            "default": "wait_for_any"
          },
          "selectors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Selectors",
            "description": "CSS selectors; the first present one wins."
          },
          "timeout_s": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Timeout S",
            "default": 15.0
          }
        },
        "type": "object",
        "required": [
          "selectors"
        ],
        "title": "WaitForAnyStep",
        "description": "Wait until ANY of the given selectors is present (whichever appears\nfirst), then continue. Avoids waiting the full timeout when e.g. an\nout-of-stock block shows up long before a (never-coming) price widget."
      },
      "WaitForStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "wait_for",
            "title": "Action",
            "default": "wait_for"
          },
          "selector": {
            "type": "string",
            "title": "Selector",
            "description": "CSS selector to wait for."
          },
          "timeout_s": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Timeout S",
            "default": 15.0
          }
        },
        "type": "object",
        "required": [
          "selector"
        ],
        "title": "WaitForStep"
      },
      "WaitForTextStep": {
        "properties": {
          "action": {
            "type": "string",
            "const": "wait_for_text",
            "title": "Action",
            "default": "wait_for_text"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Text to wait for anywhere on the page."
          },
          "timeout_s": {
            "type": "number",
            "maximum": 120.0,
            "exclusiveMinimum": 0.0,
            "title": "Timeout S",
            "default": 15.0
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "WaitForTextStep"
      }
    }
  }
}